Inventory Added
Added all the relevant files for the Inventory Component, however there's no functionality outside of being added to the player, since I do not know the native item systems and how to integrate it.

I will be linking a document in discord explaining the system so that you can integrate it from your end.